requirements.txt
pdfplumber>=0.10.0
PyMuPDF>=1.23.0
requests>=2.31.0
python-dotenv>=1.0.0
